requirements.txt
streamlit==1.32.0
nvidia-riva-client==2.14.0
nvidia-nemo-toolkit==1.23.0
openai==1.12.0
googlemaps==4.10.0
python-dotenv==1.0.0
dataclasses-json==0.6.4
pydantic==2.6.0
requests==2.31.0
uuid
typing-extensions==4.9.0
aiohttp==3.9.0
asyncio==3.4.3